--- SQLiteDemo2/src/main/java/controlador/ControladorLibro.java
@@ -1,29 +1,0 @@
-package controlador;
-
-import modelo.ModeloTablaLibro;
-import vista.VentanaLibro;
-
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
-
-public class ControladorLibro extends MouseAdapter {
-
-    private VentanaLibro view;
-    private ModeloTablaLibro modelo;
-
-    public ControladorLibro(VentanaLibro view) {
-        this.view = view;
-        this.view.getBtnCargar().addMouseListener(this);
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent e) {
-        if (e.getSource() == this.view.getBtnCargar()){
-            modelo = new ModeloTablaLibro();
-            modelo.cargarDatos();
-            this.view.getTblLibro().setModel(modelo);
-            this.view.getTblLibro().updateUI();
-        }
-    }
-}
