+++ SQLiteDemo2_/.gradle/7.5.1/checksums/sha1-checksums.bin
