+++ SQLiteDemo2_/.gradle/7.5.1/dependencies-accessors/dependencies-accessors.lock
