+++ SQLiteDemo2_/.gradle/7.5.1/fileChanges/last-build.bin
