+++ SQLiteDemo2_/.gradle/buildOutputCleanup/cache.properties
