+++ SQLiteDemo2_/build.gradle
