+++ SQLiteDemo2_/settings.gradle
