+++ SQLiteDemo2_/src/main/java/modelo/Libro.java
